requirements.txt
aiodocker==0.21.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
py-cord==2.3.2
python-dotenv==0.21.0
typing-extensions==4.3.0
yarl==1.8.1
